Upgrades to resin-io/resin-wifi-connect v3
v3 is based on NetworkManager and allows password-less access points
so we don't need to maintain our patched fork of connman anymore. It
also has an install script which simplifies our installation.

`forget-wifi-networks` clears all previous connections and reboots
the Pi. This is because as soon as this runs it will disconnect the
wifi network that's connected.

`start-wifi-connect` is a new script that Systemd will run on startup.
This runs `wifi-connect` to see if any networks can be joined. After
this if the device isn't connected to a network then it will run
`wifi-connect --clear=true` to start the AP mode. Unfortunately this
clears all existing wifi networks but this is the only way to enable
AP mode using v3.